@@ -1,7 +1,7 @@
 import React from 'react'
 import { serve } from '../../relatives/services/constant'
 import './API.css'
-import { Paper } from '@material-ui/core'
+import { Paper, Button } from '@material-ui/core'
 
 const ExampleJQuery = () => (
   <div>
@@ -82,8 +82,9 @@ class API extends React.Component<Props, State> {
             <ul>
               <li><span className="label">基础插件</span><code>{serve}/app/plugin/:repositories</code></li>
               <li><span className="label">jQuery 插件</span><code>{serve}/libs/jquery.rap.js</code>
-                <button
-                  className="btn btn-secondary btn-sm ml8"
+                <Button
+                  size="small"
+                  variant="outlined"
                   onClick={
                     e => {
                       e.preventDefault()
@@ -94,7 +95,7 @@ class API extends React.Component<Props, State> {
                   }
                 >
                   用法
-                </button>
+                </Button>
               </li>
               {this.state.showExampleJQuery && <ExampleJQuery />}
               <li><span className="label">Mock.js 插件</span><code>{serve}/libs/mock.rap.js</code></li>
