@@ -1,5 +1,5 @@
-import React from 'react'
-import { Button, makeStyles, Theme, createStyles } from '@material-ui/core'
+import React, { useState } from 'react'
+import { Button, makeStyles, Theme, createStyles, Tooltip } from '@material-ui/core'
 import LoadingButton from '../common/LoadingButton'
 import Create from '@material-ui/icons/Create'
 import KeyboardTab from '@material-ui/icons/KeyboardTab'
@@ -7,6 +7,9 @@ import Save from '@material-ui/icons/Save'
 import Cancel from '@material-ui/icons/Cancel'
 import { useSelector } from 'react-redux'
 import { RootState } from 'actions/types'
+import History from '@material-ui/icons/History'
+import HistoryLogDrawer from './HistoryLogDrawer'
+import { ENTITY_TYPE } from 'utils/consts'
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -32,28 +35,48 @@ interface Props {
   editable: boolean,
   itfId: number,
   moveInterface: any
-  handleSaveInterface: any
+  handleSaveInterfaceAndProperties: any
   handleUnlockInterface: any
   handleMoveInterface: any
-  handleLockInterface: any
+  handleEditInterface: any
 }
 
 function InterfaceEditorToolbar(props: Props) {
-  const { editable, locker, auth, repository, handleLockInterface, handleMoveInterface,
-    handleSaveInterface, handleUnlockInterface } = props
-  const isOwned = repository.owner.id === auth.id
-  const isJoined = repository.members.find((item: any) => item.id === auth.id)
+  const {
+    editable,
+    locker,
+    repository,
+    handleEditInterface,
+    handleMoveInterface,
+    handleSaveInterfaceAndProperties,
+    handleUnlockInterface,
+  } = props
+
   const loading = useSelector((state: RootState) => state.loading)
+  const [showHistory, setShowHistory] = useState(false)
 
   const classes = useStyles()
-  if (!isOwned && !isJoined) { return null }
+  if (!repository.canUserEdit) { return null }
   if (editable) {
     return (
       <div className="InterfaceEditorToolbar">
-        <LoadingButton className={classes.button} onClick={handleSaveInterface} variant="contained" color="primary" disabled={loading} label="保存">
+        <LoadingButton
+          className={classes.button}
+          onClick={handleSaveInterfaceAndProperties}
+          variant="contained"
+          color="primary"
+          disabled={loading}
+          label="保存"
+          size="small"
+        >
           <Save className={classes.rightIcon} />
         </LoadingButton>
-        <Button className={classes.button} onClick={handleUnlockInterface} variant="contained">
+        <Button
+          className={classes.button}
+          onClick={handleUnlockInterface}
+          variant="contained"
+          size="small"
+        >
           取消
           <Cancel className={classes.rightIcon} />
         </Button>
@@ -70,13 +93,47 @@ function InterfaceEditorToolbar(props: Props) {
   }
   return (
     <div className="InterfaceEditorToolbar">
-      <Button className={classes.button} onClick={handleMoveInterface} variant="contained">
-        移动 / 复制
+      <Tooltip title="查看该接口中的所有改动历史">
+        <Button
+          className={`${classes.button} guide-2`}
+          variant="contained"
+          onClick={() => setShowHistory(true)}
+          size="small"
+        >
+          历史
+        <History className={classes.rightIcon} />
+        </Button>
+      </Tooltip>
+      <Tooltip title="移动或复制该接口">
+        <Button
+          className={classes.button}
+          onClick={handleMoveInterface}
+          variant="contained"
+          size="small"
+        >
+          移动
         <KeyboardTab className={classes.rightIcon} />
-      </Button>
-      <LoadingButton className={classes.button} onClick={handleLockInterface} variant="contained" color="primary" disabled={loading} label="编辑">
-        <Create className={classes.rightIcon} />
-      </LoadingButton>
+        </Button>
+      </Tooltip>
+      <Tooltip title="点击进入编辑模式，并锁定该接口">
+        <LoadingButton
+          className={classes.button}
+          onClick={handleEditInterface}
+          variant="contained"
+          color="primary"
+          disabled={loading}
+          label="编辑"
+          size="small"
+        >
+          <Create className={classes.rightIcon} />
+        </LoadingButton>
+      </Tooltip>
+      <HistoryLogDrawer
+        open={showHistory}
+        onClose={() => setShowHistory(false)}
+        entityId={props.itfId}
+        entityType={ENTITY_TYPE.INTERFACE}
+      />
     </div>
   )
 }
