@@ -1,173 +1,214 @@
-import React, { Component } from 'react'
-import { PropTypes, connect, Mock } from '../../family'
-import { SmartTextarea } from '../utils'
-import { RootState } from 'actions/types'
-import { Button } from '@material-ui/core'
+import React from 'react'
+import { useSelector, useDispatch } from 'react-redux'
+import { YUP_MSG } from '../../family/UIConst'
+import { Formik, Field, Form } from 'formik'
+import { TextField } from 'formik-material-ui'
+import * as Yup from 'yup'
+import { Button, Theme, Dialog, DialogContent, DialogTitle, Select, MenuItem, InputLabel, FormControl } from '@material-ui/core'
+import { makeStyles } from '@material-ui/styles'
+import { SlideUp } from 'components/common/Transition'
+import { Interface, Repository, RootState, Module } from '../../actions/types'
+import { updateInterface, addInterface } from '../../actions/interface'
+import { StoreStateRouterLocationURI, push } from 'family'
 export const METHODS = ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS', 'PATCH', 'HEAD']
 export const STATUS_LIST = [200, 301, 403, 404, 500, 502, 503, 504]
-// 模拟数据
-const mockInterface =
-  process.env.NODE_ENV === 'development'
-    ? () =>
-      Mock.mock({
-        name: '接口@CTITLE(4)',
-        url: '@URL',
-        'method|1': METHODS,
-        description: '@CPARAGRAPH',
-        repositoryId: undefined,
-        moduleId: undefined,
-      })
-    : () => ({
-      name: '',
-      url: '',
-      method: 'GET',
-      description: '',
-      repositoryId: undefined,
-      moduleId: undefined,
-    })
 
-type InterfaceFormProps = any
-type InterfaceFormState = any
-class InterfaceForm extends Component<InterfaceFormProps, InterfaceFormState> {
-  static contextTypes = {
-    rmodal: PropTypes.object.isRequired,
-    onAddInterface: PropTypes.func.isRequired,
-    onUpdateInterface: PropTypes.func.isRequired,
-  }
-  constructor(props: any) {
-    super(props)
-    const itf = this.props.itf
-    this.state = itf ? { ...itf } : mockInterface()
-  }
-  render() {
-    const { rmodal } = this.context
-    return (
-      <section>
-        <div className="rmodal-header">
-          <span className="rmodal-title">{this.props.title}</span>
-        </div>
-        <form className="form-horizontal w600" onSubmit={this.handleSubmit}>
-          <div className="rmodal-body">
-            <div className="form-group row">
-              <label className="col-sm-2 control-label">名称：</label>
-              <div className="col-sm-10">
-                <input
-                  name="name"
-                  tabIndex={1}
-                  value={this.state.name}
-                  onChange={e => this.setState({ name: e.target.value })}
-                  className="form-control"
-                  placeholder="Name"
-                  spellCheck={false}
-                  autoFocus={true}
-                  required={true}
-                />
-              </div>
-            </div>
-            <div className="form-group row">
-              <label className="col-sm-2 control-label">地址：</label>
-              <div className="col-sm-10">
-                <input
-                  name="name"
-                  tabIndex={2}
-                  value={this.state.url}
-                  onChange={e => this.setState({ url: e.target.value })}
-                  className="form-control"
-                  placeholder="URI"
-                  spellCheck={false}
-                  required={true}
-                />
-              </div>
-            </div>
-            <div className="form-group row">
-              <label className="col-sm-2 control-label">类型：</label>
-              <div className="col-sm-10">
-                <select
-                  name="method"
-                  tabIndex={3}
-                  value={this.state.method}
-                  onChange={e => this.setState({ method: e.target.value })}
-                  className="form-control"
-                >
-                  {METHODS.map(method => (
-                    <option key={method} value={method}>
-                      {method}
-                    </option>
-                  ))}
-                </select>
-              </div>
-            </div>
-            <div className="form-group row">
-              <label className="col-sm-2 control-label">状态码：</label>
-              <div className="col-sm-10">
-                <select
-                  name="status"
-                  tabIndex={4}
-                  value={this.state.status}
-                  onChange={e => this.setState({ status: e.target.value })}
-                  className="form-control"
-                >
-                  {STATUS_LIST.map(status => (
-                    <option key={status} value={status}>
-                      {status}
-                    </option>
-                  ))}
-                </select>
-              </div>
-            </div>
-            <div className="form-group row">
-              <label className="col-sm-2 control-label">简介：</label>
-              <div className="col-sm-10">
-                <SmartTextarea
-                  name="description"
-                  tabIndex={5}
-                  value={this.state.description}
-                  onChange={(e: any) => this.setState({ description: e.target.value })}
-                  className="form-control"
-                  placeholder="Description"
-                  spellCheck={false}
-                  rows="5"
-                />
-              </div>
-            </div>
-          </div>
-          <div className="rmodal-footer">
-            <div className="form-group row mb0">
-              <label className="col-sm-2 control-label" />
-              <div className="col-sm-10">
-                <Button type="submit" variant="contained" color="primary" style={{marginRight: 8}}>
-                  提交
-                </Button>
-                <Button onClick={() => rmodal.close()} > 取消 </Button>
-              </div>
-            </div>
-          </div>
-        </form>
-      </section>
-    )
-  }
-  componentDidUpdate() {
-    this.context.rmodal.reposition()
-  }
-  handleSubmit = (e: any) => {
-    e.preventDefault()
-    const { onAddInterface, onUpdateInterface } = this.context
-    const { auth, repository, mod } = this.props
-    const onAddOrUpdateInterface = this.state.id ? onUpdateInterface : onAddInterface
-    const itf = Object.assign({}, this.state, {
-      creatorId: auth.id,
-      repositoryId: repository.id,
-      moduleId: mod.id,
-      lockerId: this.state.locker ? this.state.locker.id : null,
-    })
-    onAddOrUpdateInterface(itf, () => {
-      const { rmodal } = this.context
-      if (rmodal) { rmodal.resolve() }
-    })
-  };
-}
-const mapStateToProps = (state: RootState) => ({
-  auth: state.auth,
+const useStyles = makeStyles(({ spacing }: Theme) => ({
+  root: {
+  },
+  appBar: {
+    position: 'relative',
+  },
+  title: {
+    marginLeft: spacing(2),
+    flex: 1,
+  },
+  preview: {
+    marginTop: spacing(1),
+  },
+  form: {
+    minWidth: 500,
+    minHeight: 300,
+  },
+  formTitle: {
+    color: 'rgba(0, 0, 0, 0.54)',
+    fontSize: 9,
+  },
+  formItem: {
+    marginBottom: spacing(1),
+  },
+  ctl: {
+    marginTop: spacing(3),
+  },
+}))
+
+const schema = Yup.object().shape<Partial<Interface>>({
+  name: Yup.string().required(YUP_MSG.REQUIRED).max(20, YUP_MSG.MAX_LENGTH(20)),
+  description: Yup.string().max(1000, YUP_MSG.MAX_LENGTH(1000)),
 })
-const mapDispatchToProps = {}
-export default connect(mapStateToProps, mapDispatchToProps)(InterfaceForm)
+
+const FORM_STATE_INIT: Interface = {
+  id: 0,
+  name: '',
+  url: '',
+  method: 'GET',
+  description: '',
+  repositoryId: 0,
+  moduleId: 0,
+  status: 200,
+}
+
+interface Props {
+  title?: string
+  open: boolean
+  onClose: (isOk?: boolean) => void
+  itf?: Interface
+  repository?: Repository
+  mod?: Module
+}
+
+function InterfaceForm(props: Props) {
+  const auth = useSelector((state: RootState) => state.auth)
+  const { open, onClose, itf, title, repository, mod } = props
+  const classes = useStyles()
+  const dispatch = useDispatch()
+  const router = useSelector((state: RootState) => state.router)
+
+  return (
+    <Dialog
+      open={open}
+      onClose={(_event, reason) => (reason !== 'backdropClick' && onClose())}
+      TransitionComponent={SlideUp}
+    >
+      <DialogTitle>{title}</DialogTitle>
+      <DialogContent dividers={true}>
+        <div className={classes.form}>
+          <Formik
+            initialValues={{
+              ...FORM_STATE_INIT,
+              ...(itf || {}),
+            }}
+            validationSchema={schema}
+            onSubmit={values => {
+              const addOrUpdateInterface = values.id
+                ? updateInterface
+                : addInterface
+              const itf: Interface = {
+                ...values,
+                creatorId: auth.id,
+                repositoryId: repository!.id,
+                moduleId: mod!.id,
+              }
+              dispatch(
+                addOrUpdateInterface(itf, (e) => {
+                  if (e && e.id) {
+                    const href = StoreStateRouterLocationURI(router)
+                      .setSearch('itf', e.id)
+                      .href()
+                    dispatch(push(href))
+                  }
+                  onClose(true)
+                })
+              )
+            }}
+            render={({ isSubmitting, setFieldValue, values }) => {
+              return (
+                <Form>
+                  <div className="rmodal-body">
+                    <div className={classes.formItem}>
+                      <Field
+                        name="name"
+                        label="名称"
+                        component={TextField}
+                        fullWidth={true}
+                      />
+                    </div>
+                    <div className={classes.formItem}>
+                      <Field
+                        name="url"
+                        label="URL地址"
+                        component={TextField}
+                        fullWidth={true}
+                      />
+                    </div>
+                    <div className={classes.formItem}>
+                      <FormControl>
+                        <InputLabel
+                          shrink={true}
+                          htmlFor="method-label-placeholder"
+                        >
+                          类型
+                        </InputLabel>
+                        <Select
+                          value={values.method}
+                          displayEmpty={true}
+                          name="method"
+                          onChange={selected => {
+                            setFieldValue('method', selected.target.value)
+                          }}
+                        >
+                          {METHODS.map(method => (
+                            <MenuItem key={method} value={method}>
+                              {method}
+                            </MenuItem>
+                          ))}
+                        </Select>
+                      </FormControl>
+                    </div>
+                    <div className={classes.formItem}>
+                      <InputLabel
+                        shrink={true}
+                        htmlFor="method-label-placeholder"
+                      >
+                        状态码
+                      </InputLabel>
+                      <Select
+                        value={values.status}
+                        displayEmpty={true}
+                        name="status"
+                        onChange={selected => {
+                          setFieldValue('status', selected.target.value)
+                        }}
+                      >
+                        {STATUS_LIST.map(status => (
+                          <MenuItem key={status} value={status}>
+                            {status}
+                          </MenuItem>
+                        ))}
+                      </Select>
+                    </div>
+                    <div className={classes.formItem}>
+                      <Field
+                        name="description"
+                        label="说明"
+                        component={TextField}
+                        multiline={true}
+                        fullWidth={true}
+                      />
+                    </div>
+                  </div>
+                  <div className={classes.ctl}>
+                    <Button
+                      type="submit"
+                      variant="contained"
+                      color="primary"
+                      className="mr1"
+                      disabled={isSubmitting}
+                    >
+                      提交
+                    </Button>
+                    <Button onClick={() => onClose()} disabled={isSubmitting}>
+                      取消
+                    </Button>
+                  </div>
+                </Form>
+              )
+            }}
+          />
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
+
+export default InterfaceForm
