@@ -1,110 +1,105 @@
-import React, { Component } from 'react'
+import React, { useState } from 'react'
 import { connect, Link, replace, StoreStateRouterLocationURI } from '../../family'
-import { RModal, RSortable } from '../utils'
+import { RSortable } from '../utils'
 import ModuleForm from './ModuleForm'
-import { GoPencil, GoTrashcan, GoPackage } from 'react-icons/go'
-import { RootState } from 'actions/types'
-import { deleteModule } from '../../actions/module'
+import { useSelector, useDispatch } from 'react-redux'
+import { GoPackage } from 'react-icons/go'
+import { deleteModule, sortModuleList } from '../../actions/module'
+import { Module, Repository, RootState, User } from '../../actions/types'
+import { makeStyles, Theme } from '@material-ui/core'
 
-class ModuleBase extends Component<any, any> {
-  constructor(props: any) {
-    super(props)
-    this.state = { update: false }
-  }
-  render() {
-    const { auth, repository, mod, router } = this.props
-    const uri = StoreStateRouterLocationURI(router).removeSearch('itf')
-    const selectHref = uri.setSearch('mod', mod.id).href()
-    return (
-      <div className="Module clearfix">
-        <Link to={selectHref} className="name">{mod.name}</Link>
-        <div className="toolbar">
-          {/* 编辑权限：拥有者或者成员 */}
-          {repository.owner.id === auth.id || repository.members.find((item: any) => item.id === auth.id)
-            ? <span className="fake-link" onClick={() => this.setState({ update: true })}><GoPencil /></span>
-            : null
-          }
-          {repository.owner.id === auth.id || repository.members.find((item: any) => item.id === auth.id)
-            ? <span className="fake-link" onClick={e => this.handleDelete(e, mod)}><GoTrashcan /></span>
-            : null
-          }
-        </div>
-        <RModal when={this.state.update} onClose={() => this.setState({ update: false })} onResolve={this.handleUpdate}>
-          <ModuleForm title="修改模块" mod={mod} repository={repository} />
-        </RModal>
-      </div>
-    )
-  }
-  handleUpdate = () => {
-    this.props.replace(StoreStateRouterLocationURI(this.props.router).href())
-  }
-  handleDelete = (e: any, mod: any) => {
-    const { router } = this.props
-    e.preventDefault()
-    const message = `模块被删除后不可恢复，并且会删除相关的接口！\n确认继续删除『#${mod.id} ${mod.name}』吗？`
-    if (window.confirm(message)) {
-      this.props.onDeleteModule(this.props.mod.id, () => {
-        const uri = StoreStateRouterLocationURI(router)
-        const deleteHref = this.props.active ? uri.removeSearch('mod').href() : uri.href()
-        this.props.replace(deleteHref)
-      }, this.props.repository.id)
-    }
-  }
+interface ModuleBaseProps {
+  repository: Repository
+  mod: Module
+  active?: boolean
+  auth?: User
+  deleteModule: typeof deleteModule
+  replace?: typeof replace
+}
+function ModuleBase(props: ModuleBaseProps) {
+  const { mod } = props
+  const router = useSelector((state: RootState) => state.router)
+  const uri = StoreStateRouterLocationURI(router).removeSearch('itf')
+  const selectHref = uri.setSearch('mod', mod!.id.toString()).href()
+
+  return (
+    <div className="Module clearfix">
+      <Link to={selectHref} className="name">
+        {mod.name}
+      </Link>
+    </div>
+  )
 }
 
 const mapStateToModuleBaseProps = (state: any) => ({
   router: state.router,
 })
 const mapDispatchToModuleBaseProps = ({
-  onDeleteModule: deleteModule,
+  deleteModule,
   replace,
 })
 
-const Module = connect(mapStateToModuleBaseProps, mapDispatchToModuleBaseProps)(ModuleBase)
+const ModuleWrap = connect(mapStateToModuleBaseProps, mapDispatchToModuleBaseProps)(ModuleBase)
 
-class ModuleList extends Component<any, any> {
-  constructor(props: any) {
-    super(props)
-    this.state = { create: false }
-  }
-  render() {
-    const { auth, repository = {}, mods = [], mod = {} } = this.props
-    const isOwned = repository.owner.id === auth.id
-    const isJoined = repository.members.find((item: any) => item.id === auth.id)
-    return (
-      <RSortable onChange={this.handleSort} disabled={!isOwned && !isJoined}>
-        <ul className="ModuleList clearfix">
-          {mods.map((item: any) =>
-            <li key={item.id} className={item.id === mod.id ? 'active sortable' : 'sortable'} data-id={item.id}>
-              <Module key={item.id} mod={item} active={item.id === mod.id} repository={repository} auth={auth} />
-            </li>
-          )}
-          {/* 编辑权限：拥有者或者成员 */}
-          {isOwned || isJoined
-            ? <li>
-              <span className="fake-link" onClick={() => this.setState({ create: true })}>
-                <GoPackage className="fontsize-14" /> 新建模块
-              </span>
-              <RModal when={this.state.create} onClose={() => this.setState({ create: false })} onResolve={this.handleCreate}>
-                <ModuleForm title="新建模块" repository={repository} />
-              </RModal>
-            </li>
-            : null
-          }
-        </ul>
-      </RSortable>
-    )
-  }
-  handleCreate = () => {
-    const { router, replace } = this.props
-    replace(StoreStateRouterLocationURI(router).href())
-  }
-  handleSort = (_: any, sortable: any) => {
-    const { onSortModuleList } = this.context
-    onSortModuleList(sortable.toArray())
-  }
+interface ModuleListProps {
+  mods?: Module[]
+  mod?: Module
+  repository: Repository
 }
 
+const useStyles = makeStyles(({ palette }: Theme) => ({
+  li: {
+    borderColor: `${palette.primary.main} #e1e4e8 transparent #e1e4e8 !important`,
+  },
+}))
+
+function ModuleList(props: ModuleListProps) {
+  const [open, setOpen] = useState(false)
+  const dispatch = useDispatch()
+  const auth = useSelector((state: RootState) => state.auth)
+  const { repository, mods = [], mod } = props
+  const classes = useStyles()
+  const handleSort = (_: any, sortable: any) => {
+    dispatch(sortModuleList(sortable.toArray(), () => {
+      /** empty */
+    }))
+  }
+  return (
+    <RSortable onChange={handleSort} disabled={!repository.canUserEdit}>
+      <ul className="ModuleList clearfix">
+        {mods.map((item: any) => (
+          <li
+            key={item.id}
+            className={`${item.id === mod!.id ? 'active ' + classes.li : ''} sortable `}
+            data-id={item.id}
+          >
+            <ModuleWrap
+              key={item.id}
+              mod={item}
+              active={item.id === mod!.id}
+              repository={repository}
+              auth={auth}
+            />
+          </li>
+        ))}
+        {/* 编辑权限：拥有者或者成员 */}
+        {repository.canUserEdit ? (
+          <li>
+            <span onClick={() => setOpen(true)} className="g-link">
+              <GoPackage className="fontsize-14" /> 新建模块
+            </span>
+            <ModuleForm
+              title="新建模块"
+              repository={repository}
+              open={open}
+              onClose={() => setOpen(false)}
+            />
+          </li>
+        ) : null}
+      </ul>
+    </RSortable>
+  )
+}
 const mapStateToProps = (state: RootState) => ({
   auth: state.auth,
   router: state.router,
