@@ -4,15 +4,14 @@ import { connect } from 'react-redux'
 import Mock from 'mockjs'
 import SmartTextarea from '../utils/SmartTextarea'
 import { Button } from '@material-ui/core'
-
-export const TYPES = ['String', 'Number', 'Boolean', 'Object', 'Array', 'Function', 'RegExp']
+import { TYPES } from '../../utils/consts'
 
 // 模拟数据
-const mockProperty = process.env.NODE_ENV === 'development'
+export const mockProperty = process.env.NODE_ENV === 'development'
   ? () => Mock.mock({
     'scope|1': ['request', 'response'],
     name: '@WORD(6)',
-    'type|1': TYPES,
+    'type|1': ['String', 'Number', 'Boolean'],
     'value|1': ['@INT', '@FLOAT', '@TITLE', '@NAME'],
     description: '@CSENTENCE',
     parentId: -1,
@@ -38,7 +37,7 @@ class PropertyForm extends Component<any, any> {
     parent: PropTypes.object,
     repository: PropTypes.object.isRequired,
     mod: PropTypes.object.isRequired,
-    itf: PropTypes.object.isRequired,
+    interfaceId: PropTypes.number.isRequired,
   }
   static contextTypes = {
     rmodal: PropTypes.instanceOf(Component),
@@ -55,7 +54,7 @@ class PropertyForm extends Component<any, any> {
         <div className="rmodal-header">
           <span className="rmodal-title">{this.props.title}</span>
         </div>
-        <form className="form-horizontal w600" onSubmit={this.handleSubmit} >
+        <form className="form-horizontal w600" onSubmit={this.handleSubmit}>
           <div className="rmodal-body">
             <div className="form-group row" style={{}}>
               <label className="col-sm-2 control-label">名称：</label>
@@ -80,12 +79,20 @@ class PropertyForm extends Component<any, any> {
                   name="type"
                   tabIndex={2}
                   value={this.state.type}
-                  onChange={e => this.setState({ type: e.target.value })}
+                  onChange={e => {
+                    const type = e.target.value
+                    if (type === 'Null') {
+                      this.setState({ value: '' })
+                    }
+                    this.setState({ type })
+                  }}
                   className="form-control"
                 >
-                  {TYPES.map(type =>
-                    <option key={type} value={type}>{type}</option>
-                  )}
+                  {TYPES.map(type => (
+                    <option key={type} value={type}>
+                      {type}
+                    </option>
+                  ))}
                 </select>
               </div>
             </div>
@@ -137,7 +144,14 @@ class PropertyForm extends Component<any, any> {
             <div className="form-group row mb0">
               <label className="col-sm-2 control-label" />
               <div className="col-sm-10">
-                <Button type="submit" variant="contained" color="primary" style={{ marginRight: 8 }}>提交</Button>
+                <Button
+                  type="submit"
+                  variant="contained"
+                  color="primary"
+                  style={{ marginRight: 8 }}
+                >
+                  提交
+                </Button>
                 <Button onClick={() => rmodal.close()}>取消</Button>
               </div>
             </div>
@@ -151,19 +165,21 @@ class PropertyForm extends Component<any, any> {
   }
   handleSubmit = (e: any) => {
     e.preventDefault()
-    const { auth, repository, mod, itf, scope, parent = { id: -1 } } = this.props
+    const { auth, repository, mod, interfaceId, scope, parent = { id: -1 } } = this.props
     const { handleAddMemoryProperty } = this.context
     const property = Object.assign({}, this.state, {
       creatorId: auth.id,
       repositoryId: repository.id,
       moduleId: mod.id,
-      interfaceId: itf.id,
+      interfaceId,
       scope,
       parentId: parent.id,
     })
     handleAddMemoryProperty(property, () => {
       const { rmodal } = this.context
-      if (rmodal) { rmodal.resolve() }
+      if (rmodal) {
+        rmodal.resolve()
+      }
     })
   }
 }
