@@ -1,9 +1,13 @@
 import React from 'react'
 
-export default function Logo(props: {color?: string}) {
-  const color = props.color || '#FFFFFF'
+export default function Logo(_props: {color?: string}) {
   return (
-    // tslint:disable-next-line: max-line-length
-    <svg width="140px" height="34px" viewBox="0 0 263 64" version="1.1"> <g id="页面1" stroke="none" strokeWidth="1" fill={color} fillRule="evenodd"> <g id="logo"> <g id="3" transform="translate(83.000000, 40.000000)" fill={color}> <path d="M7.52,10.55 L7.52,10.55 L7.52,10.55 C7.52,9.83666667 7.31666667,9.27666667 6.91,8.87 L6.91,8.87 L6.91,8.87 C6.50333333,8.45666667 5.74333333,8.06333333 4.63,7.69 L4.63,7.69 L4.63,7.69 C3.51666667,7.31666667 2.7,6.89666667 2.18,6.43 L2.18,6.43 L2.18,6.43 C1.49333333,5.83 1.15,5.03333333 1.15,4.04 L1.15,4.04 L1.15,4.04 C1.15,3.06 1.48666667,2.26333333 2.16,1.65 L2.16,1.65 L2.16,1.65 C2.83333333,1.03666667 3.70333333,0.73 4.77,0.73 L4.77,0.73 L4.77,0.73 C5.86333333,0.73 6.75333333,1.08666667 7.44,1.8 L7.44,1.8 L7.44,1.8 C8.13333333,2.51333333 8.48,3.43333333 8.48,4.56 L8.48,4.56 L7.47,4.56 L7.47,4.56 C7.47,3.7 7.23,3.00333333 6.75,2.47 L6.75,2.47 L6.75,2.47 C6.26333333,1.94333333 5.60333333,1.68 4.77,1.68 L4.77,1.68 L4.77,1.68 C3.96333333,1.68 3.32666667,1.89666667 2.86,2.33 L2.86,2.33 L2.86,2.33 C2.39333333,2.77 2.16,3.33333333 2.16,4.02 L2.16,4.02 L2.16,4.02 C2.16,4.66666667 2.38,5.19333333 2.82,5.6 L2.82,5.6 L2.82,5.6 C3.26,6.01333333 3.96,6.38 4.92,6.7 L4.92,6.7 L4.92,6.7 C5.88,7.02 6.60666667,7.35 7.1,7.69 L7.1,7.69 L7.1,7.69 C7.59333333,8.03666667 7.95666667,8.44 8.19,8.9 L8.19,8.9 L8.19,8.9 C8.42333333,9.36 8.54,9.90666667 8.54,10.54 L8.54,10.54 L8.54,10.54 C8.54,11.54 8.20333333,12.3466667 7.53,12.96 L7.53,12.96 L7.53,12.96 C6.85,13.5733333 5.95333333,13.88 4.84,13.88 L4.84,13.88 L4.84,13.88 C4.09333333,13.88 3.41,13.7233333 2.79,13.41 L2.79,13.41 L2.79,13.41 C2.17,13.1033333 1.68666667,12.6633333 1.34,12.09 L1.34,12.09 L1.34,12.09 C0.993333333,11.51 0.82,10.8266667 0.82,10.04 L0.82,10.04 L1.82,10.04 L1.82,10.04 C1.82,10.9466667 2.09333333,11.6566667 2.64,12.17 L2.64,12.17 L2.64,12.17 C3.18,12.6833333 3.91333333,12.94 4.84,12.94 L4.84,12.94 L4.84,12.94 C5.65333333,12.94 6.30333333,12.7233333 6.79,12.29 L6.79,12.29 L6.79,12.29 C7.27666667,11.8566667 7.52,11.2766667 7.52,10.55 Z M11.57,4.19 L12.5,4.19 L12.53,5.61 L12.53,5.61 C13.1233333,4.54333333 13.9733333,4.01 15.08,4.01 L15.08,4.01 L15.08,4.01 C16.3066667,4.01 17.0966667,4.59333333 17.45,5.76 L17.45,5.76 L17.45,5.76 C17.75,5.18 18.1366667,4.74333333 18.61,4.45 L18.61,4.45 L18.61,4.45 C19.0833333,4.15666667 19.5966667,4.01 20.15,4.01 L20.15,4.01 L20.15,4.01 C21.8633333,4.01 22.74,5.11333333 22.78,7.32 L22.78,7.32 L22.78,13.7 L21.8,13.7 L21.8,7.38 L21.8,7.38 C21.7933333,6.56 21.6433333,5.95 21.35,5.55 L21.35,5.55 L21.35,5.55 C21.0566667,5.15 20.5766667,4.95 19.91,4.95 L19.91,4.95 L19.91,4.95 C19.3033333,4.95666667 18.7933333,5.19 18.38,5.65 L18.38,5.65 L18.38,5.65 C17.96,6.10333333 17.72,6.67666667 17.66,7.37 L17.66,7.37 L17.66,13.7 L16.68,13.7 L16.68,7.41 L16.68,7.41 C16.68,6.57666667 16.5266667,5.95666667 16.22,5.55 L16.22,5.55 L16.22,5.55 C15.9133333,5.15 15.4333333,4.95 14.78,4.95 L14.78,4.95 L14.78,4.95 C14.2466667,4.95 13.7866667,5.13333333 13.4,5.5 L13.4,5.5 L13.4,5.5 C13.0066667,5.86666667 12.7233333,6.39333333 12.55,7.08 L12.55,7.08 L12.55,13.7 L11.57,13.7 L11.57,4.19 Z M32.39,13.7 L31.36,13.7 L31.36,13.7 C31.2666667,13.3933333 31.2066667,12.9533333 31.18,12.38 L31.18,12.38 L31.18,12.38 C30.8866667,12.8533333 30.51,13.22 30.05,13.48 L30.05,13.48 L30.05,13.48 C29.59,13.7466667 29.09,13.88 28.55,13.88 L28.55,13.88 L28.55,13.88 C27.73,13.88 27.0933333,13.6466667 26.64,13.18 L26.64,13.18 L26.64,13.18 C26.18,12.72 25.95,12.06 25.95,11.2 L25.95,11.2 L25.95,11.2 C25.95,10.2733333 26.2666667,9.53666667 26.9,8.99 L26.9,8.99 L26.9,8.99 C27.54,8.44333333 28.42,8.16666667 29.54,8.16 L29.54,8.16 L31.17,8.16 L31.17,6.83 L31.17,6.83 C31.17,6.15666667 31.0033333,5.67 30.67,5.37 L30.67,5.37 L30.67,5.37 C30.33,5.07666667 29.8333333,4.93 29.18,4.93 L29.18,4.93 L29.18,4.93 C28.58,4.93 28.0866667,5.10666667 27.7,5.46 L27.7,5.46 L27.7,5.46 C27.3133333,5.80666667 27.12,6.26 27.12,6.82 L27.12,6.82 L26.14,6.81 L26.14,6.81 C26.14,6.03 26.4333333,5.37 27.02,4.83 L27.02,4.83 L27.02,4.83 C27.6066667,4.28333333 28.3433333,4.01 29.23,4.01 L29.23,4.01 L29.23,4.01 C30.1833333,4.01 30.9066667,4.25 31.4,4.73 L31.4,4.73 L31.4,4.73 C31.8933333,5.21666667 32.14,5.91666667 32.14,6.83 L32.14,6.83 L32.14,11.5 L32.14,11.5 C32.14,12.4333333 32.2233333,13.13 32.39,13.59 L32.39,13.59 L32.39,13.7 Z M28.67,12.93 L28.67,12.93 L28.67,12.93 C29.25,12.93 29.76,12.7833333 30.2,12.49 L30.2,12.49 L30.2,12.49 C30.64,12.19 30.9633333,11.79 31.17,11.29 L31.17,11.29 L31.17,8.98 L29.62,8.98 L29.62,8.98 C28.7333333,8.99333333 28.0633333,9.19 27.61,9.57 L27.61,9.57 L27.61,9.57 C27.1566667,9.95666667 26.93,10.4733333 26.93,11.12 L26.93,11.12 L26.93,11.12 C26.93,12.3266667 27.51,12.93 28.67,12.93 Z M39.54,4.13 L39.53,5.1 L39.53,5.1 C39.2966667,5.06 39.09,5.04 38.91,5.04 L38.91,5.04 L38.91,5.04 C38.37,5.04 37.9166667,5.22666667 37.55,5.6 L37.55,5.6 L37.55,5.6 C37.1833333,5.96666667 36.9266667,6.47666667 36.78,7.13 L36.78,7.13 L36.78,13.7 L35.8,13.7 L35.8,4.19 L36.76,4.19 L36.78,5.61 L36.78,5.61 C37.24,4.54333333 37.9666667,4.01 38.96,4.01 L38.96,4.01 L38.96,4.01 C39.2133333,4.01 39.4066667,4.05 39.54,4.13 L39.54,4.13 Z M42.73,1.76 L43.71,1.76 L43.71,4.19 L45.33,4.19 L45.33,5.07 L43.71,5.07 L43.71,11.39 L43.71,11.39 C43.71,11.91 43.7866667,12.2966667 43.94,12.55 L43.94,12.55 L43.94,12.55 C44.0933333,12.8033333 44.3533333,12.93 44.72,12.93 L44.72,12.93 L44.72,12.93 C44.88,12.93 45.1133333,12.9033333 45.42,12.85 L45.42,12.85 L45.45,13.73 L45.45,13.73 C45.23,13.83 44.91,13.88 44.49,13.88 L44.49,13.88 L44.49,13.88 C43.35,13.88 42.7633333,13.0933333 42.73,11.52 L42.73,11.52 L42.73,5.07 L41.27,5.07 L41.27,4.19 L42.73,4.19 L42.73,1.76 Z M60.64,13.7 L59.53,10.12 L54.74,10.12 L53.65,13.7 L52.6,13.7 L56.68,0.9 L57.61,0.9 L61.69,13.7 L60.64,13.7 Z M57.13,2.32 L55.03,9.18 L59.24,9.18 L57.13,2.32 Z M68.29,8.5 L65.59,8.5 L65.59,13.7 L64.59,13.7 L64.59,0.9 L68.3,0.9 L68.3,0.9 C69.4333333,0.9 70.3266667,1.24 70.98,1.92 L70.98,1.92 L70.98,1.92 C71.6266667,2.6 71.95,3.53666667 71.95,4.73 L71.95,4.73 L71.95,4.73 C71.95,5.93666667 71.6366667,6.86666667 71.01,7.52 L71.01,7.52 L71.01,7.52 C70.3766667,8.17333333 69.47,8.5 68.29,8.5 L68.29,8.5 Z M65.59,1.85 L65.59,7.56 L68.3,7.56 L68.3,7.56 C69.1466667,7.56 69.8,7.31333333 70.26,6.82 L70.26,6.82 L70.26,6.82 C70.72,6.33333333 70.95,5.64 70.95,4.74 L70.95,4.74 L70.95,4.74 C70.95,3.84666667 70.7166667,3.14333333 70.25,2.63 L70.25,2.63 L70.25,2.63 C69.7833333,2.11 69.1366667,1.85 68.31,1.85 L68.31,1.85 L65.59,1.85 Z M76.48,0.9 L76.48,13.7 L75.47,13.7 L75.47,0.9 L76.48,0.9 Z M85.48,4.19 L86.41,4.19 L86.44,5.61 L86.44,5.61 C87.0333333,4.54333333 87.8833333,4.01 88.99,4.01 L88.99,4.01 L88.99,4.01 C90.21,4.01 91,4.59333333 91.36,5.76 L91.36,5.76 L91.36,5.76 C91.66,5.18 92.0466667,4.74333333 92.52,4.45 L92.52,4.45 L92.52,4.45 C92.9866667,4.15666667 93.5,4.01 94.06,4.01 L94.06,4.01 L94.06,4.01 C95.7733333,4.01 96.65,5.11333333 96.69,7.32 L96.69,7.32 L96.69,13.7 L95.71,13.7 L95.71,7.38 L95.71,7.38 C95.6966667,6.56 95.5433333,5.95 95.25,5.55 L95.25,5.55 L95.25,5.55 C94.9566667,5.15 94.4766667,4.95 93.81,4.95 L93.81,4.95 L93.81,4.95 C93.21,4.95666667 92.7,5.19 92.28,5.65 L92.28,5.65 L92.28,5.65 C91.8666667,6.10333333 91.63,6.67666667 91.57,7.37 L91.57,7.37 L91.57,13.7 L90.59,13.7 L90.59,7.41 L90.59,7.41 C90.59,6.57666667 90.4366667,5.95666667 90.13,5.55 L90.13,5.55 L90.13,5.55 C89.8166667,5.15 89.3366667,4.95 88.69,4.95 L88.69,4.95 L88.69,4.95 C88.1566667,4.95 87.6933333,5.13333333 87.3,5.5 L87.3,5.5 L87.3,5.5 C86.9133333,5.86666667 86.6333333,6.39333333 86.46,7.08 L86.46,7.08 L86.46,13.7 L85.48,13.7 L85.48,4.19 Z M106.3,13.7 L105.26,13.7 L105.26,13.7 C105.173333,13.3933333 105.116667,12.9533333 105.09,12.38 L105.09,12.38 L105.09,12.38 C104.796667,12.8533333 104.42,13.22 103.96,13.48 L103.96,13.48 L103.96,13.48 C103.5,13.7466667 103,13.88 102.46,13.88 L102.46,13.88 L102.46,13.88 C101.64,13.88 101,13.6466667 100.54,13.18 L100.54,13.18 L100.54,13.18 C100.086667,12.72 99.86,12.06 99.86,11.2 L99.86,11.2 L99.86,11.2 C99.86,10.2733333 100.176667,9.53666667 100.81,8.99 L100.81,8.99 L100.81,8.99 C101.443333,8.44333333 102.32,8.16666667 103.44,8.16 L103.44,8.16 L105.08,8.16 L105.08,6.83 L105.08,6.83 C105.08,6.15666667 104.91,5.67 104.57,5.37 L104.57,5.37 L104.57,5.37 C104.236667,5.07666667 103.743333,4.93 103.09,4.93 L103.09,4.93 L103.09,4.93 C102.49,4.93 101.996667,5.10666667 101.61,5.46 L101.61,5.46 L101.61,5.46 C101.223333,5.80666667 101.03,6.26 101.03,6.82 L101.03,6.82 L100.05,6.81 L100.05,6.81 C100.05,6.03 100.343333,5.37 100.93,4.83 L100.93,4.83 L100.93,4.83 C101.516667,4.28333333 102.253333,4.01 103.14,4.01 L103.14,4.01 L103.14,4.01 C104.093333,4.01 104.816667,4.25 105.31,4.73 L105.31,4.73 L105.31,4.73 C105.796667,5.21666667 106.04,5.91666667 106.04,6.83 L106.04,6.83 L106.04,11.5 L106.04,11.5 C106.04,12.4333333 106.126667,13.13 106.3,13.59 L106.3,13.59 L106.3,13.7 Z M102.58,12.93 L102.58,12.93 L102.58,12.93 C103.153333,12.93 103.663333,12.7833333 104.11,12.49 L104.11,12.49 L104.11,12.49 C104.55,12.19 104.873333,11.79 105.08,11.29 L105.08,11.29 L105.08,8.98 L103.53,8.98 L103.53,8.98 C102.643333,8.99333333 101.973333,9.19 101.52,9.57 L101.52,9.57 L101.52,9.57 C101.066667,9.95666667 100.84,10.4733333 100.84,11.12 L100.84,11.12 L100.84,11.12 C100.84,12.3266667 101.42,12.93 102.58,12.93 Z M109.71,4.19 L110.64,4.19 L110.67,5.7 L110.67,5.7 C110.95,5.18666667 111.306667,4.77666667 111.74,4.47 L111.74,4.47 L111.74,4.47 C112.18,4.16333333 112.676667,4.01 113.23,4.01 L113.23,4.01 L113.23,4.01 C114.89,4.01 115.74,5.11666667 115.78,7.33 L115.78,7.33 L115.78,13.7 L114.81,13.7 L114.81,7.43 L114.81,7.43 C114.81,6.58333333 114.663333,5.95666667 114.37,5.55 L114.37,5.55 L114.37,5.55 C114.076667,5.15 113.616667,4.95 112.99,4.95 L112.99,4.95 L112.99,4.95 C112.476667,4.95 112.016667,5.14 111.61,5.52 L111.61,5.52 L111.61,5.52 C111.203333,5.90666667 110.893333,6.43333333 110.68,7.1 L110.68,7.1 L110.68,13.7 L109.71,13.7 L109.71,4.19 Z M125.44,13.7 L124.4,13.7 L124.4,13.7 C124.313333,13.3933333 124.256667,12.9533333 124.23,12.38 L124.23,12.38 L124.23,12.38 C123.936667,12.8533333 123.56,13.22 123.1,13.48 L123.1,13.48 L123.1,13.48 C122.64,13.7466667 122.14,13.88 121.6,13.88 L121.6,13.88 L121.6,13.88 C120.78,13.88 120.14,13.6466667 119.68,13.18 L119.68,13.18 L119.68,13.18 C119.226667,12.72 119,12.06 119,11.2 L119,11.2 L119,11.2 C119,10.2733333 119.316667,9.53666667 119.95,8.99 L119.95,8.99 L119.95,8.99 C120.583333,8.44333333 121.46,8.16666667 122.58,8.16 L122.58,8.16 L124.22,8.16 L124.22,6.83 L124.22,6.83 C124.22,6.15666667 124.05,5.67 123.71,5.37 L123.71,5.37 L123.71,5.37 C123.376667,5.07666667 122.883333,4.93 122.23,4.93 L122.23,4.93 L122.23,4.93 C121.63,4.93 121.136667,5.10666667 120.75,5.46 L120.75,5.46 L120.75,5.46 C120.363333,5.80666667 120.17,6.26 120.17,6.82 L120.17,6.82 L119.19,6.81 L119.19,6.81 C119.19,6.03 119.483333,5.37 120.07,4.83 L120.07,4.83 L120.07,4.83 C120.656667,4.28333333 121.39,4.01 122.27,4.01 L122.27,4.01 L122.27,4.01 C123.23,4.01 123.956667,4.25 124.45,4.73 L124.45,4.73 L124.45,4.73 C124.936667,5.21666667 125.18,5.91666667 125.18,6.83 L125.18,6.83 L125.18,11.5 L125.18,11.5 C125.18,12.4333333 125.266667,13.13 125.44,13.59 L125.44,13.59 L125.44,13.7 Z M121.72,12.93 L121.72,12.93 L121.72,12.93 C122.293333,12.93 122.803333,12.7833333 123.25,12.49 L123.25,12.49 L123.25,12.49 C123.69,12.19 124.013333,11.79 124.22,11.29 L124.22,11.29 L124.22,8.98 L122.67,8.98 L122.67,8.98 C121.776667,8.99333333 121.106667,9.19 120.66,9.57 L120.66,9.57 L120.66,9.57 C120.206667,9.95666667 119.98,10.4733333 119.98,11.12 L119.98,11.12 L119.98,11.12 C119.98,12.3266667 120.56,12.93 121.72,12.93 Z M128.5,9.53 L128.5,8.5 L128.5,8.5 C128.5,7 128.76,5.87666667 129.28,5.13 L129.28,5.13 L129.28,5.13 C129.793333,4.38333333 130.566667,4.01 131.6,4.01 L131.6,4.01 L131.6,4.01 C132.693333,4.01 133.506667,4.51333333 134.04,5.52 L134.04,5.52 L134.09,4.19 L135,4.19 L135,13.97 L135,13.97 C134.993333,15.1166667 134.72,15.99 134.18,16.59 L134.18,16.59 L134.18,16.59 C133.633333,17.19 132.843333,17.49 131.81,17.49 L131.81,17.49 L131.81,17.49 C131.29,17.49 130.763333,17.36 130.23,17.1 L130.23,17.1 L130.23,17.1 C129.703333,16.8466667 129.3,16.5366667 129.02,16.17 L129.02,16.17 L129.48,15.47 L129.48,15.47 C130.133333,16.21 130.886667,16.58 131.74,16.58 L131.74,16.58 L131.74,16.58 C132.526667,16.58 133.103333,16.37 133.47,15.95 L133.47,15.95 L133.47,15.95 C133.83,15.53 134.01,14.8866667 134.01,14.02 L134.01,14.02 L134.01,12.51 L134.01,12.51 C133.476667,13.4233333 132.67,13.88 131.59,13.88 L131.59,13.88 L131.59,13.88 C130.596667,13.88 129.84,13.5066667 129.32,12.76 L129.32,12.76 L129.32,12.76 C128.793333,12.0133333 128.52,10.9366667 128.5,9.53 L128.5,9.53 Z M129.49,8.44 L129.49,9.39 L129.49,9.39 C129.49,10.61 129.673333,11.5066667 130.04,12.08 L130.04,12.08 L130.04,12.08 C130.413333,12.6533333 130.993333,12.94 131.78,12.94 L131.78,12.94 L131.78,12.94 C132.833333,12.94 133.576667,12.44 134.01,11.44 L134.01,11.44 L134.01,6.58 L134.01,6.58 C133.576667,5.49333333 132.84,4.95 131.8,4.95 L131.8,4.95 L131.8,4.95 C131,4.95 130.416667,5.23666667 130.05,5.81 L130.05,5.81 L130.05,5.81 C129.683333,6.37666667 129.496667,7.25333333 129.49,8.44 L129.49,8.44 Z M141.74,13.88 L141.74,13.88 L141.74,13.88 C140.62,13.88 139.743333,13.5133333 139.11,12.78 L139.11,12.78 L139.11,12.78 C138.476667,12.0466667 138.153333,11.0033333 138.14,9.65 L138.14,9.65 L138.14,8.49 L138.14,8.49 C138.14,7.09666667 138.446667,6 139.06,5.2 L139.06,5.2 L139.06,5.2 C139.666667,4.40666667 140.503333,4.01 141.57,4.01 L141.57,4.01 L141.57,4.01 C142.603333,4.01 143.39,4.34666667 143.93,5.02 L143.93,5.02 L143.93,5.02 C144.47,5.7 144.746667,6.73333333 144.76,8.12 L144.76,8.12 L144.76,9.03 L139.11,9.03 L139.11,9.57 L139.11,9.57 C139.11,10.6566667 139.343333,11.4966667 139.81,12.09 L139.81,12.09 L139.81,12.09 C140.276667,12.6833333 140.933333,12.98 141.78,12.98 L141.78,12.98 L141.78,12.98 C142.72,12.98 143.48,12.6366667 144.06,11.95 L144.06,11.95 L144.57,12.6 L144.57,12.6 C143.95,13.4533333 143.006667,13.88 141.74,13.88 Z M141.57,4.93 L141.57,4.93 L141.57,4.93 C140.803333,4.93 140.213333,5.2 139.8,5.74 L139.8,5.74 L139.8,5.74 C139.386667,6.28 139.16,7.07333333 139.12,8.12 L139.12,8.12 L143.79,8.12 L143.79,7.86 L143.79,7.86 C143.743333,5.90666667 143.003333,4.93 141.57,4.93 Z M153.37,1.76 L154.35,1.76 L154.35,4.19 L155.97,4.19 L155.97,5.07 L154.35,5.07 L154.35,11.39 L154.35,11.39 C154.35,11.91 154.426667,12.2966667 154.58,12.55 L154.58,12.55 L154.58,12.55 C154.733333,12.8033333 154.993333,12.93 155.36,12.93 L155.36,12.93 L155.36,12.93 C155.513333,12.93 155.746667,12.9033333 156.06,12.85 L156.06,12.85 L156.09,13.73 L156.09,13.73 C155.863333,13.83 155.543333,13.88 155.13,13.88 L155.13,13.88 L155.13,13.88 C153.99,13.88 153.403333,13.0933333 153.37,11.52 L153.37,11.52 L153.37,5.07 L151.9,5.07 L151.9,4.19 L153.37,4.19 L153.37,1.76 Z M158.55,9.59 L158.55,8.46 L158.55,8.46 C158.55,7.08666667 158.863333,6.00333333 159.49,5.21 L159.49,5.21 L159.49,5.21 C160.11,4.41 160.956667,4.01 162.03,4.01 L162.03,4.01 L162.03,4.01 C163.103333,4.01 163.95,4.40333333 164.57,5.19 L164.57,5.19 L164.57,5.19 C165.19,5.97666667 165.503333,7.05666667 165.51,8.43 L165.51,8.43 L165.51,9.45 L165.51,9.45 C165.51,10.8433333 165.196667,11.93 164.57,12.71 L164.57,12.71 L164.57,12.71 C163.943333,13.49 163.1,13.88 162.04,13.88 L162.04,13.88 L162.04,13.88 C160.98,13.88 160.14,13.4966667 159.52,12.73 L159.52,12.73 L159.52,12.73 C158.9,11.97 158.576667,10.9233333 158.55,9.59 L158.55,9.59 Z M159.54,8.39 L159.54,9.45 L159.54,9.45 C159.54,10.5433333 159.76,11.4033333 160.2,12.03 L160.2,12.03 L160.2,12.03 C160.646667,12.6566667 161.26,12.97 162.04,12.97 L162.04,12.97 L162.04,12.97 C162.846667,12.97 163.463333,12.6666667 163.89,12.06 L163.89,12.06 L163.89,12.06 C164.31,11.46 164.523333,10.5966667 164.53,9.47 L164.53,9.47 L164.53,8.46 L164.53,8.46 C164.53,7.36666667 164.306667,6.50333333 163.86,5.87 L163.86,5.87 L163.86,5.87 C163.42,5.24333333 162.81,4.93 162.03,4.93 L162.03,4.93 L162.03,4.93 C161.27,4.93 160.666667,5.23666667 160.22,5.85 L160.22,5.85 L160.22,5.85 C159.78,6.46333333 159.553333,7.31 159.54,8.39 L159.54,8.39 Z M168.39,9.59 L168.39,8.46 L168.39,8.46 C168.39,7.08666667 168.7,6.00333333 169.32,5.21 L169.32,5.21 L169.32,5.21 C169.94,4.41 170.786667,4.01 171.86,4.01 L171.86,4.01 L171.86,4.01 C172.94,4.01 173.786667,4.40333333 174.4,5.19 L174.4,5.19 L174.4,5.19 C175.02,5.97666667 175.333333,7.05666667 175.34,8.43 L175.34,8.43 L175.34,9.45 L175.34,9.45 C175.34,10.8433333 175.026667,11.93 174.4,12.71 L174.4,12.71 L174.4,12.71 C173.78,13.49 172.94,13.88 171.88,13.88 L171.88,13.88 L171.88,13.88 C170.82,13.88 169.976667,13.4966667 169.35,12.73 L169.35,12.73 L169.35,12.73 C168.73,11.97 168.41,10.9233333 168.39,9.59 L168.39,9.59 Z M169.37,8.39 L169.37,9.45 L169.37,9.45 C169.37,10.5433333 169.593333,11.4033333 170.04,12.03 L170.04,12.03 L170.04,12.03 C170.48,12.6566667 171.093333,12.97 171.88,12.97 L171.88,12.97 L171.88,12.97 C172.68,12.97 173.293333,12.6666667 173.72,12.06 L173.72,12.06 L173.72,12.06 C174.146667,11.46 174.36,10.5966667 174.36,9.47 L174.36,9.47 L174.36,8.46 L174.36,8.46 C174.36,7.36666667 174.14,6.50333333 173.7,5.87 L173.7,5.87 L173.7,5.87 C173.253333,5.24333333 172.64,4.93 171.86,4.93 L171.86,4.93 L171.86,4.93 C171.106667,4.93 170.506667,5.23666667 170.06,5.85 L170.06,5.85 L170.06,5.85 C169.613333,6.46333333 169.383333,7.31 169.37,8.39 L169.37,8.39 Z M179.68,0.2 L179.68,13.7 L178.7,13.7 L178.7,0.2 L179.68,0.2 Z" id="形状" fillRule="nonzero"/> </g> <g id="2" transform="translate(136.000000, 5.000000)" fill={color}> <path d="M12.65,28.67 L7.83,16.36 L5.78,16.36 L5.78,27.99 L0.52,28.42 L0.52,1.42 L6.93,1.2 L6.93,1.2 C10.0233333,1.2 12.4466667,1.82 14.2,3.06 L14.2,3.06 L14.2,3.06 C15.9533333,4.29333333 16.83,6.18333333 16.83,8.73 L16.83,8.73 L16.83,8.73 C16.83,10.41 16.5,11.79 15.84,12.87 L15.84,12.87 L15.84,12.87 C15.18,13.95 14.2266667,14.85 12.98,15.57 L12.98,15.57 L15.39,21.44 L17.84,27.27 L12.65,28.67 Z M5.78,12.47 L7.58,12.47 L7.58,12.47 C8.82666667,12.47 9.80333333,12.2133333 10.51,11.7 L10.51,11.7 L10.51,11.7 C11.2166667,11.18 11.57,10.2 11.57,8.76 L11.57,8.76 L11.57,8.76 C11.57,6.53333333 10.4666667,5.42 8.26,5.42 L8.26,5.42 L5.78,5.42 L5.78,12.47 Z M23.49,28.2 L18.77,28.2 L24.46,1.53 L31.98,1.2 L37.82,27.99 L32.06,28.2 L31.08,22.41 L24.64,22.41 L23.49,28.2 Z M27.63,5.16 L25.61,18.09 L30.11,18.09 L27.63,5.16 Z M40.73,28.42 L40.73,1.2 L44.12,1.2 L44.12,1.2 C46.8266667,1.2 48.9866667,1.31666667 50.6,1.55 L50.6,1.55 L50.6,1.55 C52.2066667,1.77666667 53.6266667,2.47666667 54.86,3.65 L54.86,3.65 L54.86,3.65 C56.1,4.83 56.72,6.71333333 56.72,9.3 L56.72,9.3 L56.72,9.3 C56.72,11.6533333 56.1966667,13.4366667 55.15,14.65 L55.15,14.65 L55.15,14.65 C54.1033333,15.8633333 52.82,16.6566667 51.3,17.03 L51.3,17.03 L51.3,17.03 C49.7733333,17.3966667 48.0133333,17.58 46.02,17.58 L46.02,17.58 L46.02,27.99 L40.73,28.42 Z M46.06,13.3 L46.96,13.3 L46.96,13.3 C48.3533333,13.3 49.45,13.0133333 50.25,12.44 L50.25,12.44 L50.25,12.44 C51.0566667,11.86 51.46,10.8033333 51.46,9.27 L51.46,9.27 L51.46,9.27 C51.46,7.75666667 51.0866667,6.73666667 50.34,6.21 L50.34,6.21 L50.34,6.21 C49.6,5.68333333 48.4266667,5.42 46.82,5.42 L46.82,5.42 L46.06,5.42 L46.06,13.3 Z M59.49,28.2 L59.49,24.28 L59.49,24.28 C60.11,23.4666667 61.14,22.22 62.58,20.54 L62.58,20.54 L62.58,20.54 C64.14,18.76 65.3966667,17.24 66.35,15.98 L66.35,15.98 L66.35,15.98 C67.2966667,14.72 68.1,13.4 68.76,12.02 L68.76,12.02 L68.76,12.02 C69.42,10.64 69.75,9.29 69.75,7.97 L69.75,7.97 L69.75,7.97 C69.75,7.23 69.5166667,6.54 69.05,5.9 L69.05,5.9 L69.05,5.9 C68.5766667,5.26666667 67.8866667,4.95 66.98,4.95 L66.98,4.95 L66.98,4.95 C66.0666667,4.95 65.26,5.38666667 64.56,6.26 L64.56,6.26 L64.56,6.26 C63.8666667,7.14 63.4233333,8.19 63.23,9.41 L63.23,9.41 L58.91,8.01 L58.91,8.01 C59.15,6.71 59.63,5.50333333 60.35,4.39 L60.35,4.39 L60.35,4.39 C61.07,3.27666667 62.0066667,2.38333333 63.16,1.71 L63.16,1.71 L63.16,1.71 C64.3133333,1.03666667 65.62,0.7 67.08,0.7 L67.08,0.7 L67.08,0.7 C68.5933333,0.7 69.94,1.01333333 71.12,1.64 L71.12,1.64 L71.12,1.64 C72.2933333,2.26 73.21,3.13 73.87,4.25 L73.87,4.25 L73.87,4.25 C74.53,5.36333333 74.86,6.62666667 74.86,8.04 L74.86,8.04 L74.86,8.04 C74.86,10.6333333 73.9733333,13.2266667 72.2,15.82 L72.2,15.82 L72.2,15.82 C70.42,18.4133333 68.1266667,21.0666667 65.32,23.78 L65.32,23.78 L74.64,23.78 L74.21,28.2 L59.49,28.2 Z" id="形状" fillRule="nonzero"/> </g> <g id="1" fill={color}> <path d="M43.5292667,0.249701606 L43.3853778,0 L30.6327111,0.00739248174 L0,52.9942375 L6.36646667,63.9917861 L67.6516222,64 L74,52.9942375 L43.5292667,0.249701606 Z M6.65424444,62.4919337 L5.98248889,61.3321355 L30.9303556,18.164149 L44.7042222,41.9966888 L43.3409778,41.9966888 L30.9213111,20.5026374 L6.65424444,62.4919337 Z M27.4334444,49.5969814 L67.5652889,49.5969814 L39.4798222,1.00209197 L40.9729778,1.00127058 L69.7968,50.874238 L26.6950889,50.8750594 L27.4334444,49.5969814 Z M66.8713333,49.1961446 L27.6644889,49.1961446 L28.4028444,47.918888 L64.639,47.918888 L37.5245778,1.00291336 L39.0177333,1.00209197 L66.8713333,49.1961446 Z M63.9450444,47.5188726 L28.6338889,47.5188726 L29.3722444,46.2416161 L61.7143556,46.2407947 L35.5693333,1.00455613 L37.0624889,1.00373474 L63.9450444,47.5188726 Z M61.0204,45.8407793 L29.6032889,45.8416007 L30.3416444,44.5635227 L58.7888889,44.5635227 L33.6149111,1.00537752 L35.1072444,1.00455613 L61.0204,45.8407793 Z M58.0949333,44.1635073 L30.5735111,44.1635073 L31.2477333,42.997138 L56.0525333,42.997138 L31.7879333,1.00702029 L33.152,1.0061989 L58.0949333,44.1635073 Z M42.1874,41.9966888 L31.8249333,41.9966888 L37.0057556,33.0320726 L42.1874,41.9966888 Z M50.9835333,41.9966888 L31.0791778,7.55593773 L30.9056889,7.25613152 L2.81528889,55.8608776 L2.06788889,54.5696575 L30.8974667,4.68601204 L52.4618889,41.9975102 L50.9835333,41.9975102 L50.9835333,41.9966888 Z M30.9056889,8.05534094 L50.5222667,41.9966888 L49.0447333,41.9966888 L30.9147333,10.6254604 L3.79373333,57.5512918 L3.04633333,56.2600716 L30.9056889,8.05534094 Z M30.9147333,11.4254912 L48.5842889,41.9975102 L47.1067556,41.9975102 L30.9229556,13.9956107 L4.773,59.2417059 L4.0256,57.9504858 L30.9147333,11.4254912 Z M30.9229556,14.7948201 L46.6446667,41.9966888 L45.1654889,41.9966888 L30.9311778,17.3649396 L5.75144444,60.9329415 L5.00322222,59.6408999 L30.9229556,14.7948201 Z M32.6019333,25.4096025 L12.5306667,60.1419459 L68.7213333,60.1419459 L67.9764,61.4339874 L10.3057333,61.4339874 L31.8635778,24.1331673 L32.6019333,25.4096025 Z M13.2229778,59.7411091 L32.8329778,25.8096179 L33.5713333,27.0868745 L15.4470889,58.4490676 L69.6981333,58.4490676 L68.9523778,59.7419305 L13.2229778,59.7419305 L13.2229778,59.7411091 Z M16.1402222,58.0482308 L33.8032,27.4877113 L34.5415556,28.7649679 L18.3643333,56.7570107 L70.6741111,56.7570107 L69.9291778,58.0490522 L16.1402222,58.0490522 L16.1402222,58.0482308 Z M19.0574667,56.3569953 L34.7717778,29.1658046 L35.5101333,30.4430612 L21.2815778,55.0649537 L71.6500889,55.0649537 L70.9051556,56.3569953 L19.0574667,56.3569953 Z M21.9747111,54.6649383 L35.742,30.8430766 L36.4293778,32.0316234 L24.0245111,53.4944621 L72.5561778,53.4944621 L71.8811333,54.6649383 L21.9747111,54.6649383 Z M72.5561778,52.4940129 L25.7585778,52.4940129 L26.4632222,51.2742534 L70.4899333,51.273432 L41.4342444,1.0004492 L42.8073556,0.999627809 L72.5561778,52.4940129 Z M30.9213111,1.50724489 L54.3184667,41.9975102 L52.9223333,41.9975102 L31.0709556,4.18578744 L30.8974667,3.88598124 L1.83684444,54.1704634 L1.15604444,52.9942375 L30.9213111,1.50724489 Z M67.0727778,62.9995508 L7.52168889,62.9921583 L30.9221333,22.5027144 L31.6325333,23.7323306 L9.78608889,61.5341966 L9.61342222,61.8340028 L67.7453556,61.8340028 L67.0727778,62.9995508 Z" id="形状" fillRule="nonzero"/> </g> </g> </g> </svg>
+    <img
+      src="https://img.alicdn.com/tfs/TB1WESIN1L2gK0jSZFmXXc7iXXa-192-77.png"
+      alt="logo"
+      width="99"
+      height="40"
+      style={{ position: 'relative', top: -2 }}
+    />
   )
 }
