@@ -13,11 +13,11 @@ class OrganizationList extends Component<any, any> {
     const { name, organizations } = this.props
     if (!organizations.length) {
       return name
-        ? <div className="fontsize-14 text-center p40">没有找到匹配 <strong>{name}</strong> 的团队。</div>
-        : <div className="fontsize-14 text-center p40">没有数据。</div>
+        ? <div className="fontsize-20 text-center p50">没有找到匹配 <strong>{name}</strong> 的团队</div>
+        : <div className="fontsize-20 text-center p50">没有数据</div>
     }
     return (
-      <div className="OrganizationList">
+      <div className="OrganizationList row">
         {organizations.map((organization: any) =>
           <Organization key={organization.id} organization={organization} />
         )}
