@@ -1,70 +1,93 @@
-import React, { Component } from 'react'
-import PropTypes from 'prop-types'
-import { RParsley } from '../utils'
+import React from 'react'
 import config from '../../config'
+import {
+  Button,
+  Dialog,
+  DialogTitle,
+  DialogContent,
+} from '@material-ui/core'
+import { SlideUp } from 'components/common/Transition'
 
-class ExportPostmanForm extends Component<any, any> {
-  static contextTypes = {
-    rmodal: PropTypes.object.isRequired,
-  }
-  static propTypes = {
-    repoId: PropTypes.number,
-  }
-  rparsley: any
-  render() {
-    const { rmodal } = this.context
-    const { repoId } = this.props
-    return (
-      <section className="RepositoryForm">
-        <div className="rmodal-header">
-          <span className="rmodal-title">{this.props.title}</span>
-        </div>
-        <RParsley ref={rparsley => { this.rparsley = rparsley }}>
-          <form className="form-horizontal" onSubmit={() => false} >
-            <div className="rmodal-body">
-              <div>
-                请在Postman中点击导入（Import），选择从链接导入（Import From Link），输入如下链接即可。
-              </div>
-              <div className="alert alert-info" role="alert" style={{ margin: '8px 0' }}> {config.serve}/postman/export?id={repoId} </div>
+export default function ExportPostmanForm(props: {
+  repoId: number;
+  open: boolean;
+  onClose: () => void;
+  title: string;
+}) {
+  const { repoId, open, onClose, title } = props
+  const postmanLink = `${config.serve}/export/postman?id=${repoId}`
+  const markdownLink = `${config.serve}/export/markdown?id=${repoId}&origin=${window.location.origin}`
+  const docxLink = `${config.serve}/export/docx?id=${repoId}&origin=${window.location.origin}`
+  const rapLink =`${config.serve}/repository/get?id=${repoId}`
+
+  // const pdfLink = `${config.serve}/export/pdf?id=${repoId}&origin=${window.location.origin}`
+  return (
+    <Dialog
+      open={open}
+      onClose={() => onClose()}
+      TransitionComponent={SlideUp}
+    >
+      <DialogTitle>{title}</DialogTitle>
+      <DialogContent dividers={true}>
+        <form className="form-horizontal" onSubmit={() => false}>
+          <div className="mb5">
+            <div>
+              Postman:</div>
+            <div
+              className="alert alert-info"
+              role="alert"
+              style={{ margin: '8px 0' }}
+            >
+              <a href={postmanLink} target="_blank" rel="noopener noreferrer">
+                {postmanLink}
+              </a>
             </div>
-            <div className="rmodal-footer">
-              <div className="form-group row mb0">
-                <label className="col-sm-2 control-label" />
-                <div className="col-sm-10">
-                  <button onClick={e => { e.preventDefault(); rmodal.close() }} className="mr10 btn btn-secondary">关闭</button>
-                </div>
-              </div>
+            <div>点击以上链接下载，在 Postman 中点击导入（Import），选择从文件导入（Import File）下载的文件。</div>
+          </div>
+
+          <div>
+            <div>Markdown:</div>
+            <div
+              className="alert alert-info"
+              role="alert"
+              style={{ margin: '8px 0' }}
+            >
+              <a href={markdownLink} target="_blank" rel="noopener noreferrer">
+                {markdownLink}
+              </a>
             </div>
-          </form>
-        </RParsley>
-      </section>
-    )
-  }
-  componentDidUpdate() {
-    this.context.rmodal.reposition()
-  }
-  handleSubmit = (e: any) => {
-    e.preventDefault()
+          </div>
 
-    if (!this.rparsley.isValid()) { return }
+          <div>
+            <div>Docx:</div>
+            <div
+              className="alert alert-info"
+              role="alert"
+              style={{ margin: '8px 0' }}
+            >
+              <a href={docxLink}>{docxLink}</a>
+            </div>
+          </div>
 
-    const { onAddRepository, onUpdateRepository } = this.context
-    const onAddOrUpdateRepository = this.state.id ? onUpdateRepository : onAddRepository
-    const { organization } = this.props
-    const repository: any = {
-      ...this.state,
-      // ownerId: owner.id, // DONE 2.2 支持转移仓库
-      organizationId: organization ? organization.id : null,
-      memberIds: (this.state.members || []).map((user: any) => user.id),
-    }
-    const { owner, newOwner } = this.state
-    if (newOwner && newOwner.id !== owner.id) { repository.ownerId = newOwner.id }
-    const { rmodal } = this.context
-    rmodal.close()
-    onAddOrUpdateRepository(repository, () => {
-      if (rmodal) { rmodal.resolve() }
-    })
-  }
+          <div>
+            <div>RAP Data:</div>
+            <div
+              className="alert alert-info"
+              role="alert"
+              style={{ margin: '8px 0' }}
+            >
+              <a href={rapLink}>{rapLink}</a>
+            </div>
+            <div>用于备份，或在其它RAP2平台导入，打开后另存为保存即可。也可通过编程访问。</div>
+          </div>
+
+          <div className="mt10">
+            <Button variant="outlined" onClick={onClose}>
+              关闭
+            </Button>
+          </div>
+        </form>
+      </DialogContent>
+    </Dialog>
+  )
 }
-
-export default ExportPostmanForm
