@@ -1,193 +1,256 @@
-import React, { Component } from 'react'
-import PropTypes from 'prop-types'
-import { connect } from 'react-redux'
-import { mock } from 'mockjs'
-import { SmartTextarea, MembersInput, RParsley } from '../utils'
-import { GoInfo } from 'react-icons/go'
-import { RootState } from 'actions/types'
-import { Button } from '@material-ui/core'
-import { updateRepository, addRepository } from 'actions/repository'
+import React, { useEffect } from 'react'
+import { useSelector, useDispatch } from 'react-redux'
+import { YUP_MSG } from '../../family/UIConst'
+import { Formik, Field, Form } from 'formik'
+import { TextField } from 'formik-material-ui'
+import * as Yup from 'yup'
+import { Button, Theme, Dialog, DialogContent, DialogTitle } from '@material-ui/core'
+import { makeStyles } from '@material-ui/styles'
+import { RepositoryFormData, RootState, Repository } from '../../actions/types'
+import UserList from '../common/UserList'
+import Select from '../common/Select'
+import AccountService from '../../relatives/services/Account'
+import * as _ from 'lodash'
+import { updateRepository, addRepository } from '../../actions/repository'
+import { fetchOwnedOrganizationList, fetchJoinedOrganizationList } from '../../actions/organization'
+import { refresh } from '../../actions/common'
+import { SlideUp } from 'components/common/Transition'
 
-// 模拟数据
-// DONE 2.1 各种表单的初始值混乱，待重构
-const mockRepository = process.env.NODE_ENV === 'development'
-  ? () => mock({
-    name: '仓库@CTITLE(6)',
-    description: '@CPARAGRAPH',
-    members: [],
-    ownerId: undefined,
-    organizationId: undefined,
-    collaboratorIds: [],
-  })
-  : () => ({
-    name: '',
-    description: '',
-    members: [],
-    ownerId: undefined,
-    organizationId: undefined,
-    collaboratorIds: [],
-  })
+const useStyles = makeStyles(({ spacing }: Theme) => ({
+  root: {},
+  appBar: {
+    position: 'relative',
+  },
+  title: {
+    marginLeft: spacing(2),
+    flex: 1,
+  },
+  preview: {
+    marginTop: spacing(1),
+  },
+  form: {
+    minWidth: 500,
+    minHeight: 300,
+  },
+  formTitle: {
+    color: 'rgba(0, 0, 0, 0.54)',
+    fontSize: 9,
+  },
+  formItem: {
+    marginBottom: spacing(1),
+  },
+  ctl: {
+    marginTop: spacing(3),
+  },
+}))
 
-class RepositoryForm extends Component<any, any> {
-  static contextTypes = {
-    rmodal: PropTypes.object.isRequired,
-  }
-  static propTypes = {
-    auth: PropTypes.object.isRequired,
-    organization: PropTypes.object,
-    repository: PropTypes.object,
-  }
-  rparsley: any
-  constructor(props: any) {
-    super(props)
-    const { repository } = props
-    this.state = repository ? {
-      ...repository,
-      collaboratorIds: repository.collaborators.map((item: any) => item.id),
-      newOwner: repository.owner,
-    } : mockRepository()
-  }
-  render() {
-    const { rmodal } = this.context
-    const { auth } = this.props
-    return (
-      <section className="RepositoryForm">
-        <div className="rmodal-header">
-          <span className="rmodal-title">{this.props.title}</span>
-        </div>
-        <RParsley ref={rparsley => { this.rparsley = rparsley }}>
-          <form className="form-horizontal" onSubmit={this.handleSubmit} >
-            <div className="rmodal-body">
-              {this.state.id &&
-                <div className="form-group row">
-                  <label className="col-sm-2 control-label">拥有者：</label>
-                  <div className="col-sm-10">
-                    {this.state.owner && (this.state.owner.id === auth.id)
-                      ? <MembersInput
-                        value={this.state.newOwner ? [this.state.newOwner] : []}
-                        limit={1}
-                        onChange={(users: any) => this.setState({ newOwner: users[0] })}
-                      />
-                      : <div className="pt7 pl9">{this.state.owner.fullname}</div>
-                    }
-                  </div>
-                </div>
-              }
-              {/* <div className='form-group row'>
-                <label className='col-sm-2 control-label'>权限</label>
-                <div className='col-sm-10'>
-                  <RadioList data={FORM.RADIO_LIST_DATA_VISIBILITY} curVal={this.state.visibility} name='visibility'
-                    onChange={visibility => this.setState({ visibility })} />
-                </div>
-              </div> */}
-              <div className="form-group row">
-                <label className="col-sm-2 control-label">名称：</label>
-                <div className="col-sm-10">
-                  <input
-                    name="name"
-                    value={this.state.name}
-                    onChange={e => this.setState({ name: e.target.value })}
-                    className="form-control"
-                    placeholder="Name"
-                    spellCheck={false}
-                    autoComplete="off"
-                    autoFocus={true}
-                    required={true}
-                    data-parsley-trigger="change keyup"
-                    data-parsley-maxlength="256"
-                  />{/* w280 */}
-                </div>
-              </div>
-              <div className="form-group row">
-                <label className="col-sm-2 control-label">简介：</label>
-                <div className="col-sm-10">
-                  <SmartTextarea
-                    name="description"
-                    value={this.state.description}
-                    onChange={(e: any) => this.setState({ description: e.target.value })}
-                    className="form-control"
-                    placeholder="Description"
-                    spellCheck={false}
-                    rows="5"
-                    data-parsley-trigger="change keyup"
-                    data-parsley-maxlength="1024"
-                  />
-                </div>
-              </div>
-              <div className="form-group row">
-                <label className="col-sm-2 control-label">成员：</label>
-                <div className="col-sm-10">
-                  <MembersInput value={this.state.members} onChange={(members: any) => this.setState({ members })} />
-                </div>
-              </div>
-              <div className="form-group row">
-                {/* DONE 2.1 帮助信息：仓库 ID 用逗号分隔，例如 1,2,3 */}
-                <label className="col-sm-2 control-label">协同仓库：</label>
-                <div className="col-sm-10">
-                  {/* TODO 2.2 CollaboratorsInput */}
-                  <input
-                    name="name"
-                    value={this.state.collaboratorIds.join(',')}
-                    onChange={e => this.setState({ collaboratorIds: e.target.value.split(',') })}
-                    className="form-control"
-                    placeholder="Collaborator Ids"
-                    spellCheck={false}
-                    autoComplete="off"
-                  />
-                  <div className="mt6 color-9">
-                    <GoInfo className="mr3" />
-                    指定与哪些仓库共享接口，仓库 ID 之间用逗号分隔，例如 <code>1,2,3</code>。</div>
-                </div>
-              </div>
-            </div>
-            <div className="rmodal-footer">
-              <div className="form-group row mb0">
-                <label className="col-sm-2 control-label" />
-                <div className="col-sm-10">
-                  <Button type="submit" variant="contained" color="primary" style={{ marginRight: 8 }}>提交</Button>
-                  <Button onClick={e => { e.preventDefault(); rmodal.close() }} variant="contained">取消</Button>
-                </div>
-              </div>
-            </div>
-          </form>
-        </RParsley>
-      </section>
-    )
-  }
-  componentDidUpdate() {
-    this.context.rmodal.reposition()
-  }
-  handleSubmit = (e: any) => {
-    e.preventDefault()
-    if (!this.rparsley.isValid()) { return }
-    const { addRepository, updateRepository } = this.props
-    const onAddOrUpdateRepository = this.state.id ? updateRepository : addRepository
-    const { organization } = this.props
-    const repository: any = {
-      ...this.state,
-      organizationId: organization ? organization.id : null,
-      memberIds: (this.state.members || []).map((user: any) => user.id),
-    }
-    const { owner, newOwner } = this.state
-    if (newOwner && newOwner.id !== owner.id) { repository.ownerId = newOwner.id }
-    const { rmodal } = this.context
-    rmodal.close()
-    onAddOrUpdateRepository(repository, () => {
-      if (rmodal) { rmodal.resolve() }
-    })
-  }
+const schema = Yup.object().shape<Partial<Repository>>({
+  name: Yup.string()
+    .required(YUP_MSG.REQUIRED)
+    .max(20, YUP_MSG.MAX_LENGTH(20)),
+  description: Yup.string().max(1000, YUP_MSG.MAX_LENGTH(1000)),
+})
+
+const FORM_STATE_INIT: RepositoryFormData = {
+  id: 0,
+  name: '',
+  description: '',
+  members: [],
+  organizationId: undefined,
+  collaborators: [],
+  collaboratorIdstring: '',
+}
+interface Props {
+  title?: string
+  open: boolean
+  onClose: (isOk?: boolean) => void
+  repository?: Repository
+  organizationId?: number
 }
 
-// 容器组件
-const mapStateToProps = (state: RootState) => ({
-  auth: state.auth,
-})
-const mapDispatchToProps = ({
-  updateRepository,
-  addRepository,
-})
+function RepositoryForm(props: Props) {
+  const { open, onClose, title, organizationId } = props
+  let repository = props.repository as RepositoryFormData
+  if (repository) {
+    repository.collaboratorIdstring = repository.collaborators!.map(x => { return x.id }).join(',')
+  }
+  const auth = useSelector((state: RootState) => state.auth)
+  const organizations = useSelector((state: RootState) => {
+    return _.uniqBy([...state.ownedOrganizations.data, ...state.joinedOrganizations.data], 'id')
+  }).map(org => ({
+    label: org.name,
+    value: org.id,
+  }))
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(RepositoryForm)
+  const classes = useStyles()
+  const dispatch = useDispatch()
+
+  useEffect(() => {
+    dispatch(fetchJoinedOrganizationList())
+    dispatch(fetchOwnedOrganizationList())
+  }, [dispatch])
+
+  if (repository) {
+    repository = { ...FORM_STATE_INIT, ...repository }
+    repository.collaboratorIdstring = repository
+      .collaborators!.map(x => {
+        return x.id
+      })
+      .join(',')
+  } else {
+    repository = { ...FORM_STATE_INIT }
+    if (organizationId !== undefined) {
+      repository.organizationId = organizationId
+    }
+  }
+
+  return (
+    <Dialog
+      open={open}
+      onClose={(_event, reason) => reason !== 'backdropClick' && onClose()}
+      TransitionComponent={SlideUp}
+    >
+      <DialogTitle>{title}</DialogTitle>
+      <DialogContent dividers={true}>
+        <div className={classes.form}>
+          <Formik
+            initialValues={repository}
+            validationSchema={schema}
+            onSubmit={values => {
+              const addOrUpdateRepository = values.id ? updateRepository : addRepository
+              const repository: RepositoryFormData = {
+                ...values,
+                memberIds: (values.members || []).map((user: any) => user.id),
+                collaboratorIds: (values.collaboratorIdstring || '').split(','),
+              }
+              const { owner, newOwner } = values
+              if (newOwner && newOwner.id !== owner!.id) {
+                repository.ownerId = newOwner.id
+              }
+              dispatch(
+                addOrUpdateRepository(repository, () => {
+                  dispatch(refresh())
+                  onClose(true)
+                }),
+              )
+            }}
+            render={({ isSubmitting, setFieldValue, values }) => {
+              function loadUserOptions(
+                input: string,
+              ): Promise<{ label: string; value: number }[]> {
+                return new Promise(async resolve => {
+                  const users = await AccountService.fetchUserList({ name: input })
+                  const options = _.differenceWith(users.data, values.members || [], _.isEqual)
+                  resolve(
+                    options.map(x => ({
+                      label: `${x.fullname} ${x.empId || x.email}`,
+                      value: x.id,
+                    })),
+                  )
+                })
+              }
+
+              const selectOrganization = organizations.find(
+                (x: any) => x.value === values.organizationId,
+              )
+              return (
+                <Form>
+                  <div className="rmodal-body">
+                    {values.id > 0 && (
+                      <div className={classes.formItem}>
+                        <div className={classes.formTitle}>拥有者</div>
+
+                        {values.owner && values.owner.id === auth.id ? (
+                          <UserList
+                            isMulti={false}
+                            loadOptions={loadUserOptions}
+                            selected={
+                              values.owner
+                                ? [
+                                    {
+                                      label: values.owner.fullname,
+                                      value: values.owner.id,
+                                    },
+                                  ]
+                                : []
+                            }
+                            onChange={(users: any) => setFieldValue('newOwner', users[0])}
+                          />
+                        ) : (
+                          <div className="pt7 pl9">{values.owner!.fullname}</div>
+                        )}
+                      </div>
+                    )}
+                    <div className={classes.formItem}>
+                      <Field name="name" label="仓库名称" component={TextField} fullWidth={true} />
+                    </div>
+                    <div className={classes.formItem}>
+                      <Field
+                        name="description"
+                        label="说明(多行，支持Markdown)"
+                        multiline={true}
+                        component={TextField}
+                        fullWidth={true}
+                        rowsMax={8}
+                      />
+                    </div>
+                    <div className={classes.formItem}>
+                      <div className={classes.formTitle}>成员</div>
+                      <UserList
+                        isMulti={true}
+                        loadOptions={loadUserOptions}
+                        selected={values.members!.map(x => ({
+                          label: x.fullname,
+                          value: x.id,
+                        }))}
+                        onChange={selected => setFieldValue('members', selected)}
+                      />
+                    </div>
+                    <div className={classes.formItem}>
+                      <div className={classes.formTitle}>团队</div>
+                      <Select
+                        isMulti={false}
+                        isClearable={true}
+                        options={organizations}
+                        value={selectOrganization}
+                        onChange={val => {
+                          if (!Array.isArray(val)) {
+                            setFieldValue('organizationId', val && val.value)
+                          }
+                        }}
+                      />
+                    </div>
+                    <div className={classes.formItem}>
+                      <Field
+                        name="collaboratorIdstring"
+                        label="协同仓库ID"
+                        component={TextField}
+                        fullWidth={true}
+                      />
+                    </div>
+                  </div>
+                  <div className={classes.ctl}>
+                    <Button
+                      type="submit"
+                      variant="contained"
+                      color="primary"
+                      className="mr1"
+                      disabled={isSubmitting}
+                    >
+                      提交
+                    </Button>
+                    <Button onClick={() => onClose()} disabled={isSubmitting}>
+                      取消
+                    </Button>
+                  </div>
+                </Form>
+              )
+            }}
+          />
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
+
+export default RepositoryForm
