@@ -11,6 +11,8 @@ import Tree from './Tree'
 import RSortable from './RSortable'
 import RParsley from './RParsley'
 import RChart from './RChart'
+import CopyToClipboard from './CopyToClipboard'
+import CustomScroll from './CustomScroll'
 
 export {
   NoMatch,
@@ -25,5 +27,7 @@ export {
   Tree,
   RSortable,
   RParsley,
-  RChart
+  RChart,
+  CopyToClipboard,
+  CustomScroll,
 }
