@@ -1,5 +1,6 @@
+
 const config: IConfig = {
-  serve: 'http://127.0.0.1:8080',
+  serve: `http://${window.location.hostname}:8080`,
   keys: ['some secret hurr'],
   session: {
     key: 'koa:sess',
