@@ -1,5 +1,3 @@
-import 'bootstrap'
-import 'bootstrap/dist/css/bootstrap.min.css'
 import './assets/index.css'
 import 'animate.css'
 
@@ -16,7 +14,13 @@ import EditorRelative from './relatives/EditorRelative'
 import Spin from './components/utils/Spin'
 import * as AccountAction from './actions/account'
 import AccountService from './relatives/services/Account'
+import { CACHE_KEY } from 'utils/consts'
+import { AnyAction } from 'redux'
 
+if (process.env.NODE_ENV !== 'production') {
+  const whyDidYouRender = require('@welldone-software/why-did-you-render/dist/no-classes-transpile/umd/whyDidYouRender.min.js')
+  whyDidYouRender(React)
+}
 // 渲染整站开屏动画
 function* renderOpeningScreenAdvertising() {
   yield new Promise((resolve) => {
@@ -39,10 +43,11 @@ function* authenticate() {
   const auth = yield call(AccountService.fetchLoginInfo)
   if (auth) {
     yield put(AccountAction.fetchLoginInfoSucceeded(auth))
+    yield put(AccountAction.fetchUserSettings([CACHE_KEY.THEME_ID]) as AnyAction)
   } else {
     const { pathname, search, hash } = window.location
     // const uri = URI(pathname + search + hash)
-    if (pathname.indexOf('/account/login') > -1) {
+    if (pathname.indexOf('/account/login') > -1 || pathname.indexOf('/account/resetpwd') > -1 ) {
       yield Promise.resolve()
       return
     }
