@@ -1,13 +1,53 @@
-import { call, put, select } from 'redux-saga/effects'
+import { call, put, select, take } from 'redux-saga/effects'
 import * as AccountAction from '../actions/account'
 import * as CommonAction from '../actions/common'
 import AccountService from './services/Account'
 import { StoreStateRouterLocationURI, replace, push } from '../family'
 import { RootState } from '../actions/types'
 import { showMessage, MSG_TYPE } from 'actions/common'
+import { THEME_TEMPLATE_KEY } from 'components/account/ThemeChangeOverlay'
+import { CHANGE_THEME, DoUpdateUserSettingAction, updateUserSetting, UPDATE_USER_SETTING_SUCCESS, DO_UPDATE_USER_SETTING, UPDATE_ACCOUNT_SUCCESS, UPDATE_ACCOUNT_FAILURE } from '../actions/account'
+import { AnyAction } from 'redux'
+import { createCommonDoActionSaga } from './effects/commonSagas'
 
 const relatives = {
   reducers: {
+    userSettings(state: RootState['userSettings'] = {}, action: any) {
+      switch (action.type) {
+        case AccountAction.FETCH_USER_SETTINGS_SUCCESS: {
+          if (action.payload.isOk) {
+            return {
+              ...state,
+              ...action.payload.data,
+            }
+          }
+          break
+        }
+      }
+      return state
+    },
+    userSettingsIsUpdating(state: boolean = false, action: any) {
+      switch (action.type) {
+        case AccountAction.UPDATE_USER_SETTING_REQUEST:
+        case AccountAction.FETCH_USER_SETTINGS_REQUEST:
+          return true
+        case AccountAction.UPDATE_USER_SETTING_FAILURE:
+        case AccountAction.UPDATE_USER_SETTING_SUCCESS:
+        case AccountAction.FETCH_USER_SETTINGS_FAILURE:
+        case AccountAction.FETCH_USER_SETTINGS_SUCCESS:
+          return false
+      }
+      return state
+    },
+    themeId(state: THEME_TEMPLATE_KEY = THEME_TEMPLATE_KEY.INDIGO, action: any) {
+      switch (action.type) {
+        case CHANGE_THEME:
+          return action.payload
+
+        default:
+          return state
+      }
+    },
     loading(state: boolean = false, action: any) {
       switch (action.type) {
         case 'INTERFACE_LOCK':
@@ -29,6 +69,8 @@ const relatives = {
     },
     auth(state: any = {}, action: any) {
       switch (action.type) {
+        case AccountAction.findpwdSucceeded().type:
+        case AccountAction.findpwdFailed('').type:
         case AccountAction.loginSucceeded({}).type:
         case AccountAction.fetchLoginInfoSucceeded({}).type:
           return action.user && action.user.id ? action.user : {}
@@ -43,6 +85,8 @@ const relatives = {
     },
     user(state: any = {}, action: any) {
       switch (action.type) {
+        case AccountAction.findpwdSucceeded().type:
+        case AccountAction.findpwdFailed('').type:
         case AccountAction.loginSucceeded({}).type:
         case AccountAction.fetchLoginInfoSucceeded({}).type:
           return action.user && action.user.id ? action.user : {}
@@ -83,6 +127,18 @@ const relatives = {
     },
   },
   sagas: {
+    *[AccountAction.DO_FETCH_USER_SETTINGS](action: AccountAction.DoFetchUserSettingsAction) {
+      const { keys, cb } = action.payload
+      yield put(AccountAction.fetchUserSettings(keys) as AnyAction)
+      const resultAction = yield take(AccountAction.FETCH_USER_SETTINGS_SUCCESS)
+      cb && cb(true, resultAction.payload)
+    },
+    *[AccountAction.FETCH_USER_SETTINGS_SUCCESS](action: any) {
+      const themeId = action.payload?.data?.THEME_ID
+      if (themeId) {
+        yield put(AccountAction.changeTheme(themeId))
+      }
+    },
     *[CommonAction.refresh().type]() {
       const router = yield select((state: RootState) => state.router)
       const uri = StoreStateRouterLocationURI(router)
@@ -101,7 +157,9 @@ const relatives = {
     *[AccountAction.addUser(null, null).type](action: any) {
       try {
         const user = yield call(AccountService.addUser, action.user)
+        let isOk = false
         if (user && user.id) {
+          isOk = true
           yield put(AccountAction.addUserSucceeded(user))
           try {
             const user = yield call(AccountService.fetchLoginInfo)
@@ -113,10 +171,11 @@ const relatives = {
           }
           yield put(push('/'))
         } else {
+          yield put(showMessage(`注册失败：${user.errMsg}`, MSG_TYPE.ERROR))
           yield put(AccountAction.addUserFailed('注册失败'))
         }
         if (action.onResolved) {
-          action.onResolved()
+          action.onResolved(isOk)
         }
       } catch (e) {
         yield put(AccountAction.addUserFailed(e.message))
@@ -177,6 +236,42 @@ const relatives = {
         yield put(AccountAction.fetchUserListFailed(e.message))
       }
     },
+    *[AccountAction.findpwd({}, () => {/** empty */ }).type](action: any) {
+      try {
+        const result = yield call(AccountService.findpwd, action.user)
+        if (result.errMsg) {
+          throw new Error(result.errMsg)
+        }
+        yield put(AccountAction.findpwdSucceeded())
+        if (action.onResolved) { action.onResolved() }
+      } catch (e) {
+        yield put(showMessage(e.message, MSG_TYPE.WARNING))
+        yield put(AccountAction.findpwdFailed(e.message))
+      }
+    },
+    *[AccountAction.resetpwd({}, () => {/** empty */ }).type](action: any) {
+      try {
+        const result = yield call(AccountService.resetpwd, action.user)
+        if (result.errMsg) {
+          throw new Error(result.errMsg)
+        }
+        yield put(AccountAction.resetpwdSucceeded())
+        if (action.onResolved) { action.onResolved() }
+      } catch (e) {
+        yield put(showMessage(e.message, MSG_TYPE.WARNING))
+        yield put(AccountAction.resetpwdFailed(e.message))
+      }
+    },
+    *[DO_UPDATE_USER_SETTING](action: DoUpdateUserSettingAction) {
+      const { key, value, cb } = action.payload
+      yield put(updateUserSetting(key, value) as AnyAction)
+      const opAction = yield take(UPDATE_USER_SETTING_SUCCESS)
+      cb && cb(opAction.payload.isOk)
+    },
+    *[AccountAction.DO_UPDATE_ACCOUNT](action: ReturnType<typeof AccountAction.doUpdateAccount>) {
+      console.log(`saga run`)
+      yield createCommonDoActionSaga(AccountAction.updateAccount, UPDATE_ACCOUNT_SUCCESS, UPDATE_ACCOUNT_FAILURE)(action)
+    }
   },
   listeners: {
     '/account': [AccountAction.fetchUserList],
