@@ -1,19 +1,19 @@
 import {
   call,
-  put
+  put,
+  select
 } from 'redux-saga/effects'
 import * as ModuleAction from '../../actions/module'
 import * as RepositoryAction from '../../actions/repository'
 import EditorService from '../services/Editor'
+import { RootState } from 'actions/types'
+import { replace } from 'family'
 
 export function* addModule(action: any) {
   try {
     const module = yield call(EditorService.addModule, action.module)
     yield put(ModuleAction.addModuleSucceeded(module))
-    yield put(RepositoryAction.fetchRepository({
-      id: action.module.repositoryId,
-      repository: undefined,
-    }))
+    yield put(RepositoryAction.refreshRepository())
     if (action.onResolved) { action.onResolved() }
   } catch (e) {
     console.error(e.message)
@@ -45,14 +45,28 @@ export function* updateModule(action: any) {
     if (action.onRejected) { action.onRejected() }
   }
 }
+export function* moveModule(action: any) {
+  try {
+    const params = action.params
+    yield call(EditorService.moveModule, params)
+    const router = yield select((state: RootState) => state.router)
+    const { pathname, hash, query } = router.location
+    yield put(replace(pathname + hash + `?id=${query.id}`))
+    yield put(RepositoryAction.refreshRepository())
+    action.onResolved && action.onResolved()
+  } catch (e) {
+    console.error(e.message)
+    action.onRejected && action.onRejected()
+  }
+}
 export function* deleteModule(action: any) {
   try {
     const count = yield call(EditorService.deleteModule, action.id)
     yield put(ModuleAction.deleteModuleSucceeded(count))
-    yield put(RepositoryAction.fetchRepository({
-      id: action.repoId,
-      repository: undefined,
-    }))
+    const router = yield select((state: RootState) => state.router)
+    const { pathname, hash, query } = router.location
+    yield put(replace(pathname + hash + `?id=${query.id}`))
+    yield put(RepositoryAction.refreshRepository())
     if (action.onResolved) { action.onResolved() }
   } catch (e) {
     console.error(e.message)
@@ -63,7 +77,7 @@ export function* deleteModule(action: any) {
 export function* sortModuleList(action: any) {
   try {
     const count = yield call(EditorService.sortModuleList, action.ids)
-    yield put(ModuleAction.sortModuleListSucceeded(count))
+    yield put(ModuleAction.sortModuleListSucceeded(count, action.ids))
     if (action.onResolved) { action.onResolved() }
   } catch (e) {
     console.error(e.message)
