@@ -29,7 +29,8 @@ export default {
       ...CREDENTIALS,
       method: 'POST',
       body: JSON.stringify({
-        email, password,
+        email,
+        password,
       }),
       headers: {
         'Content-Type': 'application/json',
@@ -54,7 +55,11 @@ export default {
     email,
     password,
     captcha,
-  }: { email: string, password: string, captcha: string }) {
+  }: {
+    email: string;
+    password: string;
+    captcha: string;
+  }) {
     return fetch(`${serve}/account/login`, {
       ...CREDENTIALS,
       method: 'POST',
@@ -80,7 +85,7 @@ export default {
   },
   // 获取用户总数
   fetchUserCount() {
-    return fetch(`${serve}/account/count`)
+    return fetch(`${serve}/account/count`, { ...CREDENTIALS })
       .then(res => res.json())
       .then(json => json.data)
   },
@@ -89,26 +94,81 @@ export default {
     name = '',
     cursor = 1,
     limit = 100,
-  }: { name?: string, cursor?: number, limit?: number } = {}) {
-    return fetch(`${serve}/account/list?name=${name}&cursor=${cursor}&limit=${limit}`)
-      .then(res => res.json())
+  }: { name?: string; cursor?: number; limit?: number } = {}) {
+    return fetch(
+      `${serve}/account/list?name=${name}&cursor=${cursor}&limit=${limit}`,
+      {
+        ...CREDENTIALS,
+      }
+    ).then(res => res.json())
     // .then(json => json.data)
   },
   // 根据 id 删除指定用户
   deleteUser(id: number) {
-    return fetch(`${serve}/account/remove?id=${id}`)
+    return fetch(`${serve}/account/remove?id=${id}`, {
+      ...CREDENTIALS,
+    })
       .then(res => res.json())
       .then(json => json.data)
   },
   // 获取用户列表
-  fetchLogList({
-    cursor = 1,
-    limit = 100,
-  }: { cursor: number, limit: number }) {
+  fetchLogList({ cursor = 1, limit = 100 }: { cursor: number; limit: number }) {
     return fetch(`${serve}/account/logger?cursor=${cursor}&limit=${limit}`, {
       ...CREDENTIALS,
+    }).then(res => res.json())
+    // .then(json => json.data)
+  },
+  // 发送重置密码激活邮件
+  findpwd({
+    email,
+    captcha,
+  }: {
+    email: string;
+    captcha: string;
+  }) {
+    return fetch(`${serve}/account/findpwd`, {
+      ...CREDENTIALS,
+      method: 'POST',
+      body: JSON.stringify({
+        email,
+        captcha,
+      }),
+      headers: {
+        'Content-Type': 'application/json',
+      },
     })
       .then(res => res.json())
-    // .then(json => json.data)
+      .then(json => json.data)
+  },
+  // 通过邮件链接重置密码
+  resetpwd({
+    email,
+    code,
+    token,
+    password,
+    captcha,
+  }: {
+    email: string;
+    code: string;
+    token: string;
+    password: string;
+    captcha: string;
+  }) {
+    return fetch(`${serve}/account/findpwd/reset`, {
+      ...CREDENTIALS,
+      method: `POST`,
+      body: JSON.stringify({
+        code,
+        email,
+        captcha,
+        token,
+        password,
+      }),
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    })
+      .then(res => res.json())
+      .then(json => json.data)
   },
 }
